backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
python-pptx>=0.6.23
opencv-python-headless>=4.8.0
Pillow>=10.0.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
PyMuPDF
scikit-learn>=1.0.0
structlog>=24.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.25.0
